Internal ASN1 reorganisation. The use of (ASN1_VALUE *) in the free and encoding routines causes problems with non pointer types (such as ASN1_BOOLEAN) and have to be treated as special cases.
Changed this to use (ASN1_VALUE **) throughout which makes
results in (ASN1_BOOLEAN *) which is a pointer type and can be
treated in a similar way to others.

This is in preparation for customised primitive types. This
would allow (for example) an ASN1_INTEGER to be represented
as a BIGNUM directly or an int or long. This will be needed
for some of the remaining ASN1 code relating to public key
representations.